requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
httpx==0.24.1
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.15
asyncpg==0.27.0
alembic==1.11.1
python-dotenv==1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart==0.0.6
email-validator==2.1.0.post1